TSI: color processing implemented
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@16951 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/TSI/TSICam.cpp
@@ -541,7 +541,7 @@ int TsiCam::SnapImage()
    if (color)
    {
       MM::MMTime start = GetCurrentMMTime();
-      MM::MMTime timeout(TSI_MAX_EXPOSURE_MS / 1000, 0); // upper limit on the exposure
+      MM::MMTime timeout((long)(TSI_MAX_EXPOSURE_MS / 1000), 0L); // upper limit on the exposure
       int err(TSI_NO_ERROR);
 
       bool computeWhiteBalanceCoefficients = false;
@@ -590,7 +590,7 @@ int TsiCam::SnapImage()
    {
       // grayscale image snap
       MM::MMTime start = GetCurrentMMTime();
-      MM::MMTime timeout(TSI_MAX_EXPOSURE_MS / 1000, 0); // upper limit on the exposure
+      MM::MMTime timeout((long)(TSI_MAX_EXPOSURE_MS / 1000), 0L); // upper limit on the exposure
       TsiImage* tsiImg = 0;
       do
       {
